requirements.txt
streamlit
streamlit-component-lib
openai
python-dotenv
markdown
